🚑	긴급 수정 : 방에 시드머니 Integer 엔티티 정의 되어있어서 Long 변경

+++ backend/src/main/java/com/kmbbj/backend/games/service/gamebalance/GameBalanceServiceImpl.java
@@ -30,7 +30,7 @@ public class GameBalanceServiceImpl implements GameBalanceService {
     @Transactional
     public List<GameBalance> createGameBalance(Game game) {
         // 방의 시작 시드머니 가져옴
-        Integer seedMoney = game.getRoom().getStartSeedMoney();
+        Long seedMoney = Long.valueOf((game.getRoom().getStartSeedMoney()));
 
 
         // 실제로 게임을 플레이한 사용자 리스트 가져오기
@@ -57,5 +57,8 @@ public class GameBalanceServiceImpl implements GameBalanceService {
         }
         return gameBalances;
     }
+
+
+
 }
